Split test_metrics.py (683 lines) and test_github.py (503 lines) into focused files
test_metrics.py -> test_cycle_time.py, test_pr_metrics.py,
  test_pickup_review_time.py, test_bot_login.py
test_github.py -> test_github_fetch.py, test_github_pagination.py

All 6 new files are well under the 400-line limit. No test logic changed.
